Dynamic string decryptor for unknown protection
Implemented a basic string decryption engine that can decrypt strings in
an assembly with unknown protection based on only a metadata token of
the string decryption method.

How to? Pass metadata token with the -dtok param. E.g -dtok:0000001

It only has support for a single decryption method for now.